+++ sphaera/diffeq/_impl/solvers.py
@@ -25,7 +25,7 @@ class AdaptiveStepsizeODESolver(object):
     def integrate(self, t):
         _assert_increasing(t)
         solution = [self.y0]
-        t = t.to(self.y0[0].device, torch.float32)
+        t = t.to(self.y0[0].device, torch.float64)
         self.before_integrate(t)
         for i in range(1, len(t)):
             y = self.advance(t[i])
